refactor: rename api folder
